requirements.txt
fastapi
uvicorn
pandas
openai
langchain
langchain_community
langchain_openai
python-dotenv
pydantic
streamlit
uvicorn